requirements.txt
altgraph==0.17.4
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pygame==2.6.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
PySimpleGUI==5.0.8.3
python-dateutil==2.9.0.post0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
urllib3==2.5.0
